Create xor() - script to convert data.py
In this example, the xor_string_with_integer function takes an input string and XORs each character with the integer 13. The ord() function is used to get the Unicode code point of a character, the XOR operation is performed, and then the result is converted back to a character using chr(). 

You can use this function to XOR any string with the integer 13 or any other value.